requirements.txt
streamlit
youtube_dl
SpeechRecognition
pydub
yt_dlp
youtube-transcript-api
google-generativeai
pytube
langdetect
python-dotenv
python-multipart